+++ AspNetCore/AT.Common.AspNetCore.Publish/Extensions/StartupExtensions.cs
@@ -1,3 +1,4 @@
+using System.Text.Json.Serialization;
 using System.Text.RegularExpressions;
 using Arbeidstilsynet.Common.AspNetCore.Extensions.CrossCutting;
 using Microsoft.AspNetCore.Builder;
@@ -38,7 +39,15 @@ public static partial class StartupExtensions
         configureMvcAction ??= options => options.Filters.Add<RequestValidationFilter>();
 
         services.Configure<RouteOptions>(options => options.LowercaseUrls = true);
-        services.AddControllers(configureMvcAction);
+        services
+            .AddControllers(configureMvcAction)
+            .ConfigureApplicationPartManager(manager =>
+            {
+                manager.FeatureProviders.Add(new CustomControllerFeatureProvider());
+            })
+            .AddJsonOptions(options =>
+                options.JsonSerializerOptions.Converters.Add(new JsonStringEnumConverter())
+            );
         services.AddProblemDetails(configureProblemDetailsAction);
         var healthChecksBuilder = services.AddHealthChecks();